+++ app/views/articles/show.html.haml
@@ -34,3 +34,6 @@
   = link_to new_article_comment_path(@article) do
     .btn-secondary
       コメントを投稿
+
+
+= javascript_pack_tag 'article'
